+++ ValidationApp/src/Inventory/InventoryItem.hpp
@@ -1,19 +1,19 @@
 #pragma once
 
-#include <memory>
+#include "Engine/Delta.hpp"
 #include "../Items/Item.hpp"
 
 class InventoryItem
 {
    public:
-	InventoryItem(std::shared_ptr<Item> item, int amount);
-	void AddAmount(int amount);
-	void LowerAmount(int amount);
-	int GetAmount() const;
+	InventoryItem(std::shared_ptr<Item> item, Uint8 amount);
+	void AddAmount(Uint8 amount);
+	void LowerAmount(Uint8 amount);
+	Uint8 GetAmount() const;
 
 	Item* GetItem() const;
 
    private:
-	int _amount;
+	Uint8 _amount;
 	std::shared_ptr<Item> _item;
 };
